Fixed Duplication bug, added config
When using /DropWhenFull, the block would be added to the inventory anyway and drop, duplicating the block being destroyed. Further, when turned on, it could not be turned off. Both are now fixed.

I added a config.yml file that is created when the plugin is enabled that allows the user to change if DropWhenFull is on or off by default.